Second attempt at conversion between sage.Integer and integer

+++ sboxU/utils.py
@@ -3,4 +3,4 @@
 from sboxu_cpp import oplus_cpp
 
 def oplus(x, y):
-    return oplus_cpp(int(x), int(y))
+    return oplus_cpp(long(x), long(y))
